[#950] 시간 표시 포맷터를 CommonPresentation으로 승격 — 편집 조작은 EventDetailScene 잔류
- SelectTimeText·SelectedTime — EventTime→표시문자열 변환 코어(케이스 4개·isAllDay·init(_:_:)·startDate)와 이를 지탱하는 extension Date(yearText·dateText·timeText·isSameDay·components) 전체가 CommonPresentation으로 이동
- SelectedTime(EventDetailScene) — isValid·eventTime(_:)·toggleIsAllDay(_:)를 extension SelectedTime으로 재작성해 편집 조작만 잔류, Optional 확장(periodStartChanged·periodEndTimeChanged·removePeriodEndTime)은 그대로 유지
- SelectedTime·SelectTimeText를 참조하던 EventDetailScene 14개 파일(Sources 8·Tests 6)에 import CommonPresentation 추가
