+++ Presentations/EventDetailScene/Sources/DoneTodoDetail/DoneTodoDetailViewModel.swift
@@ -14,6 +14,7 @@ import Prelude
 import Optics
 import Domain
 import Scenes
+import CommonPresentation
 
 
 // MARK: - DoneTodoDetailViewModel
