+++ Presentations/EventDetailScene/Sources/Models/SelectedTime.swift
@@ -10,65 +10,13 @@ import Prelude
 import Optics
 import Domain
 import Extensions
+import CommonPresentation
 
 
 // MARK: - selectedTime
 
-struct SelectTimeText: Equatable {
-    var year: String?
-    let day: String
-    var time: String?
-    let date: Date
-    
-    init(_ timeStamp: TimeInterval, _ timeZone: TimeZone, withoutTime: Bool = false) {
-        let date = Date(timeIntervalSince1970: timeStamp)
-        let isSameYear = Date().components(timeZone).0 == date.components(timeZone).0
-        self.year = isSameYear ? nil : date.yearText(at: timeZone)
-        self.day = date.dateText(at: timeZone)
-        self.time = withoutTime ? nil : date.timeText(at: timeZone)
-        self.date = date
-    }
-    
-    static func == (_ lhs: Self, _ rhs: Self) -> Bool {
-        return lhs.year == rhs.year && lhs.day == rhs.day && lhs.time == rhs.time
-    }
-}
+extension SelectedTime {
 
-enum SelectedTime: Equatable {
-    case at(SelectTimeText)
-    case period(SelectTimeText, SelectTimeText)
-    case singleAllDay(SelectTimeText)
-    case alldayPeriod(SelectTimeText, SelectTimeText)
-    
-    var isAllDay: Bool {
-        switch self {
-        case .singleAllDay, .alldayPeriod: return true
-        default: return false
-        }
-    }
-    
-    init(_ time: EventTime, _ timeZone: TimeZone) {
-        switch time {
-        case .at(let timeStamp):
-            self = .at(
-                .init(timeStamp, timeZone)
-            )
-            
-        case .period(let range):
-            self = .period(
-                .init(range.lowerBound, timeZone), .init(range.upperBound, timeZone)
-            )
-            
-        case .allDay:
-            let range = time.rangeWithShifttingifNeed(on: timeZone)
-            let isSameDay = Date(timeIntervalSince1970: range.lowerBound)
-                .isSameDay(Date(timeIntervalSince1970: range.upperBound), at: timeZone)
-            self = isSameDay
-            ? .singleAllDay(.init(range.lowerBound, timeZone, withoutTime: true))
-            : .alldayPeriod(.init(range.lowerBound, timeZone, withoutTime: true), .init(range.upperBound, timeZone, withoutTime: true))
-        }
-    }
-    
     var isValid: Bool {
         switch self {
         case .period(let start, let end): return start.date < end.date
@@ -77,26 +25,17 @@ enum SelectedTime: Equatable {
         }
     }
 
-    var startDate: Date {
-        switch self {
-        case .at(let time): return time.date
-        case .singleAllDay(let time): return time.date
-        case .period(let start, _): return start.date
-        case .alldayPeriod(let start, _): return start.date
-        }
-    }
-    
     func eventTime(_ timeZone: TimeZone) -> EventTime? {
         let calendar = Calendar(identifier: .gregorian) |> \.timeZone .~ timeZone
         let secondsFromGMT = timeZone.secondsFromGMT() |> TimeInterval.init
         switch self {
         case .at(let time):
             return .at(time.date.timeIntervalSince1970)
-            
+
         case .period(let start, let end):
             guard start.date < end.date else { return nil }
             return .period(start.date.timeIntervalSince1970..<end.date.timeIntervalSince1970)
-            
+
         case .singleAllDay(let time):
             guard let end = calendar.endOfDay(for: time.date) else { return nil }
             let start = calendar.startOfDay(for: time.date)
@@ -114,7 +53,7 @@ enum SelectedTime: Equatable {
             )
         }
     }
-    
+
     func toggleIsAllDay(_ timeZone: TimeZone) -> SelectedTime? {
         let calendar = Calendar(identifier: .gregorian) |> \.timeZone .~ timeZone
         switch self {
@@ -141,11 +80,11 @@ enum SelectedTime: Equatable {
 }
 
 extension Optional where Wrapped == SelectedTime {
-    
+
     func periodStartChanged(_ date: Date, _ timeZone: TimeZone) -> SelectedTime {
-    
+
         let timeText = SelectTimeText(date.timeIntervalSince1970, timeZone)
-        
+
         return switch self {
             case .none, .at: .at(timeText)
             case .period(_, let end): .period(timeText, end)
@@ -156,7 +95,7 @@ extension Optional where Wrapped == SelectedTime {
             case .alldayPeriod(_, let end): .alldayPeriod(timeText |> \.time .~ nil, end)
         }
     }
-    
+
     func periodEndTimeChanged(_ date: Date, _ timeZone: TimeZone) -> SelectedTime? {
         let timeText = SelectTimeText(date.timeIntervalSince1970, timeZone)
         return switch self {
@@ -168,51 +107,12 @@ extension Optional where Wrapped == SelectedTime {
             case .alldayPeriod(let start, _): .alldayPeriod(start, timeText |> \.time .~ nil)
         }
     }
-    
+
     func removePeriodEndTime(_ timeZone: TimeZone) -> SelectedTime? {
         return switch self {
         case .period(let start, _): .at(start)
         case .alldayPeriod(let start, _): .singleAllDay(start)
         default: nil
         }
-    }
-}
-
-
-extension Date {
-    
-    func yearText(at timeZone: TimeZone) -> String {
-        let dateForm = DateFormatter()
-        dateForm.timeZone = timeZone
-        dateForm.dateFormat = R.String.DateForm.yyyy
-        return dateForm.string(from: self)
-    }
-    
-    func dateText(at timeZone: TimeZone) -> String {
-        let dateForm = DateFormatter()
-        dateForm.timeZone = timeZone
-        dateForm.dateFormat = R.String.DateForm.mmmDdE
-        return dateForm.string(from: self)
-    }
-    
-    func timeText(at timeZone: TimeZone) -> String {
-        let timeForm = DateFormatter()
-        timeForm.timeZone = timeZone
-        timeForm.dateFormat = R.String.DateForm.hhMm
-        return timeForm.string(from: self)
-    }
-    
-    func isSameDay(_ other: Date, at timeZone: TimeZone) -> Bool {
-        let lhsCompos = self.components(timeZone)
-        let rhsCompos = other.components(timeZone)
-        return lhsCompos.0 == rhsCompos.0
-            && lhsCompos.1 == rhsCompos.1
-            && lhsCompos.2 == rhsCompos.2
-    }
-    
-    func components(_ timeZone: TimeZone) -> (Int?, Int?, Int?) {
-        let calendar = Calendar(identifier: .gregorian) |> \.timeZone .~ timeZone
-        let compos = calendar.dateComponents([.year, .month, .day], from: self)
-        return (compos.year, compos.month, compos.day)
     }
 }
