+++ Presentations/EventDetailScene/Sources/ViewModels/AddEventViewModel.swift
@@ -14,6 +14,7 @@ import Optics
 import Domain
 import Extensions
 import Scenes
+import CommonPresentation
 
 
 // MARK: - AddEventViewModelImple
