+++ Presentations/EventDetailScene/Tests/AddEventViewModelImpleTests.swift
@@ -11,6 +11,7 @@ import Prelude
 import Optics
 import Domain
 import Scenes
+import CommonPresentation
 import UnitTestHelpKit
 import TestDoubles
