+++ Presentations/EventDetailScene/Tests/GoogleCalendarEventDetailViewModelImpleTests.swift
@@ -12,6 +12,7 @@ import Prelude
 import Optics
 import Domain
 import Scenes
+import CommonPresentation
 import Extensions
 import UnitTestHelpKit
 import TestDoubles
